network: fix network packets corruption
Before this patch, the skb timestamp was used as the key for a map
shared between the cgroup/skp program and the kprobe program
responsible for calling it. The reason why this was needed is that
there is absolute no helper available inside the cgroup/skb context in
kernels below 5.6. The technique used here is to pre-process all the
events in a kprobe, before the ingress/egress cgroup/skb programs, and
simply submit the result (together with the packet headers/data) from
the ingress/egress programs.

Unfortunately, preemption is enabled in between the kprobe and ingress
or egress programs. This means that the map can't be a per cpu map,
since the task can migrate cpus, and the key that indexes the
pre-processed work has to be known inside the cgroup/skb program,
which is very hard without having bpf helpers to use things such as
host tid, and similar keys.

Unfortunately, only skb timestamp as the key is also not enough. There
are times that the timestamp is zeroed and it would cause conflicts
and make tracee to generate errors such as:

  {"L":"ERROR","T":"2023-02-10T16:31:57.883-0300","M":"tracee
      encountered an error","error":"failed to derive event 2006:
      resource record length exceeds data"}

  {"L":"ERROR","T":"2023-02-10T16:31:57.883-0300","M":"tracee
      encountered an error","error":"failed to derive event 2007:
      resource record length exceeds data"}

  {"L":"ERROR","T":"2023-02-10T16:31:57.883-0300","M":"tracee
      encountered an error","error":"failed to derive event 2008:
      resource record length exceeds data"}

There were other issues as well like:

1. Logic should ignore IPv4/IPv6 dual stack (unsupported for now)
2. Logic should ignore IPIP and other tunneled protocols (for now)
3. Logic should ignore other protocols (than TCP/UDP/ICMP/ICMPv6)

The only values shared in between the kprobe and the cgroup/skb
programs are either the skb timestamp, skb data pointer value OR the
ctx pointer itself (the kprobe skb argument and ctx argument from the
cgroup/skb program, although the skb program pointer points to the VM
memory).

And there is more problem: the "sk_buff" structure available to the
cgroup/skb program can't be used as a bpf map key or value if it
contains arithmetic (like creating a hash using 2 or 3 fields from the
"sk_buff" passed as the ctx argument to the cgroup/skb program).

So, instead of relying only on the skb timestamp, this patch makes the
kprobe logic to parse sb_buff data, read the L3 header contents and
use the IP/IPv6 header fields to index the event context temporarily
in a map until the cgroup/skb program runs.

Throughput results with net_packet_ipv4 filtering (the worst case),
with "--output none", seem satisfactory:

[  5] local 127.0.0.1 port 33388 connected to 127.0.0.1 port 5201
[ ID] Interval           Transfer     Bitrate         Retr  Cwnd
[  5]   0.00-1.00   sec  1.25 GBytes  10.8 Gbits/sec    0    895 KBytes
[  5]   1.00-2.00   sec  1.23 GBytes  10.6 Gbits/sec    0   1.75 MBytes
[  5]   2.00-3.00   sec  1.26 GBytes  10.9 Gbits/sec    0   1.75 MBytes
[  5]   3.00-4.00   sec  1.25 GBytes  10.8 Gbits/sec    0   1.75 MBytes
[  5]   4.00-5.00   sec  1.31 GBytes  11.2 Gbits/sec    0   1.75 MBytes
[  5]   5.00-6.00   sec   998 MBytes  8.38 Gbits/sec    0   1.75 MBytes
[  5]   6.00-7.00   sec  1.08 GBytes  9.26 Gbits/sec    0   1.75 MBytes
[  5]   7.00-8.00   sec  1.01 GBytes  8.68 Gbits/sec    0   1.75 MBytes
[  5]   8.00-9.00   sec  1.33 GBytes  11.5 Gbits/sec    0   1.75 MBytes
[  5]   9.00-10.00  sec  1.09 GBytes  9.32 Gbits/sec    0   1.81 MBytes
- - - - - - - - - - - - - - - - - - - - - - - - -
[ ID] Interval           Transfer     Bitrate         Retr
[  5]   0.00-10.00  sec  11.8 GBytes  10.1 Gbits/sec    0   sender
[  5]   0.00-10.00  sec  11.8 GBytes  10.1 Gbits/sec        receiver

This patch also removes the kretprobe after the cgroup/skb program
runs, since it is not needed anymore.